# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ngaturi Digital Invitation API
  version: 0.1.0
  description: API contract source of truth. Dates are YYYY-MM-DD, timestamps are RFC 3339 with timezone, and timezones use IANA names.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      operationId: register
      summary: Register an owner account
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } }
      responses:
        '201':
          description: Account and session created
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthSuccess' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Start an owner session
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200':
          description: Session created
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthSuccess' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshToken
      summary: Refresh an owner session
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RefreshRequest' } } }
      responses:
        '200':
          description: Session refreshed
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthSuccess' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: End an owner session
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Session ended
          content: { application/json: { schema: { $ref: '#/components/schemas/EmptySuccess' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Unauthorized:
      description: Authentication required or invalid
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    Conflict:
      description: State or uniqueness conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, details, request_id]
          properties:
            code: { type: string, examples: [INVITATION_NOT_FOUND] }
            message: { type: string, examples: ['Undangan tidak ditemukan'] }
            details: { type: [object, array, string, number, boolean, 'null'] }
            request_id: { type: string }
    EmptySuccess:
      type: object
      required: [success, data]
      properties: { success: { const: true }, data: { type: object, maxProperties: 0 } }
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, format: password, minLength: 12, maxLength: 128 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties: { refresh_token: { type: string, minLength: 20 } }
    AuthSuccess:
      type: object
      required: [success, data]
      properties:
        success: { const: true }
        data:
          type: object
          required: [access_token, refresh_token, token_type, expires_at]
          properties:
            access_token: { type: string }
            refresh_token: { type: string }
            token_type: { const: Bearer }
            expires_at: { type: string, format: date-time }
...